backend/requirements.txt
geographiclib==2.0
geopy==2.3.0
foliumasgiref==3.6.0
backports.zoneinfo==0.2.1
branca==0.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
folium==0.14.0
geographiclib==2.0
geopy==2.3.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
pytz==2023.2
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.15
